src/bili_live_helper/requirements.txt
mcdreforged
bilibili-api-python